[TASk] authorized class content

+++ app/GraphQL/Mutations/ClassContent/DeleteClassContent.php
@@ -2,9 +2,13 @@
 
 namespace App\GraphQL\Mutations\ClassContent;
 
+use App\Exceptions\ClassContentException;
 use GraphQL\Type\Definition\ResolveInfo;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use App\Models\ClassContent;
+use App\Models\Classes;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class DeleteClassContent
 {
@@ -19,6 +23,17 @@ class DeleteClassContent
      */
     public function __invoke($rootValue, array $args, GraphQLContext $context, ResolveInfo $resolveInfo)
     {
+
+
+      $class_id = ClassContent::findOrFail($args['id'])->class_id;
+      $teacher_id = Classes::findOrFail($class_id)->teacher_id;
+      $user_id = User::findOrFail($teacher_id);
+       if(Auth::id() !== $user_id){
+        throw new ClassContentException(
+               'Unantorize to delete since you are not owner of this class content',
+               'Unanthorize to delete.'
+           );
+        }
         $classContent = ClassContent::findOrFail($args['id']);
         $classContent->delete();
         //delete previous media
